+++ chat-client/src/client/mynahUi.test.ts
@@ -37,6 +37,7 @@ describe('MynahUI', () => {
             sourceLinkClick: sinon.stub(),
             infoLinkClick: sinon.stub(),
             uiReady: sinon.stub(),
+            disclaimerAcknowledged: sinon.stub(),
         }
 
         messager = new Messager(outboundChatApi)
@@ -46,7 +47,7 @@ describe('MynahUI', () => {
         const tabFactory = new TabFactory({})
         createTabStub = sinon.stub(tabFactory, 'createTab')
         createTabStub.returns({})
-        const mynahUiResult = createMynahUi(messager, tabFactory)
+        const mynahUiResult = createMynahUi(messager, tabFactory, true)
         mynahUi = mynahUiResult[0]
         inboundChatApi = mynahUiResult[1]
         getSelectedTabIdStub = sinon.stub(mynahUi, 'getSelectedTabId')
@@ -117,7 +118,7 @@ describe('MynahUI', () => {
             getSelectedTabIdStub.returns(undefined)
             inboundChatApi.sendGenericCommand({ genericCommand, selection, tabId, triggerType })
 
-            sinon.assert.calledOnceWithExactly(createTabStub, false)
+            sinon.assert.calledOnceWithExactly(createTabStub, false, false)
             sinon.assert.calledThrice(updateStoreSpy)
         })
 
@@ -133,11 +134,11 @@ describe('MynahUI', () => {
             getSelectedTabIdStub.returns(tabId)
             inboundChatApi.sendGenericCommand({ genericCommand, selection, tabId, triggerType })
 
-            sinon.assert.calledOnceWithExactly(createTabStub, false)
+            sinon.assert.calledOnceWithExactly(createTabStub, false, false)
             sinon.assert.calledThrice(updateStoreSpy)
         })
 
-        it('should not create a new tab if one exits already', () => {
+        it('should not create a new tab if one exists already', () => {
             createTabStub.resetHistory()
             const genericCommand = 'Explain'
             const selection = 'const x = 5;'